test: add analytics service tests
- extract analytics service tests
